rename metada-service to application-registry in tests
