+++ tests/application-registry-tests/Gopkg.lock
@@ -69,12 +69,12 @@
     "components/application-operator/pkg/client/clientset/versioned",
     "components/application-operator/pkg/client/clientset/versioned/scheme",
     "components/application-operator/pkg/client/clientset/versioned/typed/applicationconnector/v1alpha1",
-    "components/metadata-service/pkg/apis/istio/v1alpha2",
-    "components/metadata-service/pkg/client/clientset/versioned",
-    "components/metadata-service/pkg/client/clientset/versioned/scheme",
-    "components/metadata-service/pkg/client/clientset/versioned/typed/istio/v1alpha2"
+    "components/application-registry/pkg/apis/istio/v1alpha2",
+    "components/application-registry/pkg/client/clientset/versioned",
+    "components/application-registry/pkg/client/clientset/versioned/scheme",
+    "components/application-registry/pkg/client/clientset/versioned/typed/istio/v1alpha2"
   ]
-  revision = "d6ba814ef8b42b4c7c65f478e51f46f01387e1df"
+  revision = "d37318b3a84115bd9ee64d3c3130ef657b8a81a5"
 
 [[projects]]
   name = "github.com/modern-go/concurrent"
@@ -306,6 +306,6 @@
 [solve-meta]
   analyzer-name = "dep"
   analyzer-version = 1
-  inputs-digest = "163ec51fe838e89e1becade9287b39a739a637ef2d286ec163645453cd786c21"
+  inputs-digest = "22d6116eb8b7c96b0c721ca9f1832b652602f66dc325a962577511cee81d1f2d"
   solver-name = "gps-cdcl"
   solver-version = 1
