+++ tests/application-registry-tests/test/testkit/k8sresources_check.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	application "github.com/kyma-project/kyma/components/application-operator/pkg/apis/applicationconnector/v1alpha1"
-	istio "github.com/kyma-project/kyma/components/metadata-service/pkg/apis/istio/v1alpha2"
+	istio "github.com/kyma-project/kyma/components/application-registry/pkg/apis/istio/v1alpha2"
 	"github.com/stretchr/testify/require"
 	v1core "k8s.io/api/core/v1"
 )
