+++ tests/application-registry-tests/test/testkit/k8sresources_client.go
@@ -2,12 +2,13 @@ package testkit
 
 import (
 	"fmt"
+
 	"github.com/kyma-project/kyma/components/application-operator/pkg/apis/applicationconnector/v1alpha1"
 	"github.com/kyma-project/kyma/components/application-operator/pkg/client/clientset/versioned"
-	"github.com/kyma-project/kyma/components/metadata-service/pkg/apis/istio/v1alpha2"
-	istioclient "github.com/kyma-project/kyma/components/metadata-service/pkg/client/clientset/versioned"
+	"github.com/kyma-project/kyma/components/application-registry/pkg/apis/istio/v1alpha2"
+	istioclient "github.com/kyma-project/kyma/components/application-registry/pkg/client/clientset/versioned"
 	v1core "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1"
+	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/rand"
 
 	"k8s.io/client-go/kubernetes"
